Combine run reader with run implementation
