+++ include/faint/Run.h
@@ -1,6 +1,7 @@
-#ifndef RUN_CONFIG_H
-#define RUN_CONFIG_H
+#ifndef FAINT_RUN_H
+#define FAINT_RUN_H
 
+#include <map>
 #include <string>
 
 #include <nlohmann/json.hpp>
@@ -26,6 +27,26 @@ struct Run {
     void validate() const;
 };
 
-}
+class RunReader {
+public:
+  void add(Run rc);
+
+  const Run& get(const std::string& beam, const std::string& period) const;
+
+  const std::map<std::string, Run>& all() const noexcept;
+
+  static RunReader from_json(const nlohmann::json& data);
+
+  static RunReader from_file(const std::string& path);
+
+  static RunReader read(const nlohmann::json& data) { return from_json(data); }
+
+  static RunReader read(const std::string& path) { return from_file(path); }
+
+private:
+  std::map<std::string, Run> configs_;
+};
+
+}  // namespace faint
 
 #endif
